Compute the random unit interval in double, and keep it below 1
RAND_MAX is 2^31-1 wherever rand() is a 31-bit generator, and float cannot
represent that value: converting it rounds up to 2^31. So RAND_MAX + 1.0f
does not produce the bound it appears to, and clang refuses the silent change
of value outright, which broke the macOS build. Do the arithmetic in double,
where the +1 is exact, then narrow the finished quotient. The divisor works
out to the same 2^31 either way, so the numbers callers see are unchanged.

That left a second problem the warning had drawn attention to. The quotient is
always below 1 in double, but float carries only 24 bits of mantissa, so the
largest draw narrows to exactly 1.0f, and the helper never honoured the [0,1)
range its callers assume. Reverbtron builds an index as RND() * span and reads
data[] at it, so that draw indexes one past the end -- with span 2000 it asks
for element 2000 of 2000. It needs rand() to return exactly RAND_MAX, so it is
rare rather than harmless, and it cannot happen on MinGW at all, where
RAND_MAX is 32767 and the quotient stays clear of the bound. Clamp to the
largest float below 1.0f.

RND1 differed from RND only in casting RAND_MAX before the addition, which
reached the same value by the same rounding; it now defers to RND rather than
restating it.

+++ extra/rakverb_lib.cpp
@@ -36,7 +36,11 @@ inline FILE* portable_fopen(const char* filename, const char* mode)
 #endif
 }
 
-#define RND (rand() / ((RAND_MAX) + 1.0f))
+// RAND_MAX is 2^31-1 wherever rand() is a 31-bit generator, and float cannot
+// represent that: converting it rounds up to 2^31, so RAND_MAX + 1.0f never
+// lands on the intended bound and clang rejects the silent change of value.
+// Divide in double, where the +1 is exact, and narrow the finished quotient.
+#define RND (static_cast<float>(rand() / (static_cast<double>(RAND_MAX) + 1.0)))
 
 } // anonymous namespace
